Allow for custom messages
This allows for the client to send custom messages to the server using the `client.message` API and the server to respond using the `onMessage` plugin callback option.

These are akin to raw access to the channel and do not provide any higher-level hapi features such as auth, request-based logging, etc.